+++ CHANGELOG.md
@@ -52,7 +52,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### State Machine Breaking
 
-
+- (recovery) [#27](https://github.com/zkMeLabs/evmos/pull/27) Remove `x/recovery` module.
 
 ### API Breaking